requirement.txt
Django==4.2.0
crewai==0.28.0
google-generativeai==0.3.0
djangorestframework==3.14.0
python-dotenv==1.0.0
channels==4.0.0
redis==5.0.0
celery==5.3.0
pandas==2.2.2
numpy==1.24.0